+++ modules/auras.lua
@@ -334,7 +334,7 @@ local function ShouldShowBuff(unit, auraData, frameType)
             local filterOnlyMe = BetterBlizzFramesDB["targetBuffFilterOnlyMe"] and isTargetFriendly and (caster == "player" or (caster == "pet" and UnitIsUnit(caster, "pet")))
             if shouldBlacklist then
                 local isInBlacklist, allowMine = isInBlacklist(spellName, spellId)
-                if isInBlacklist and not allowMine then return end
+                if isInBlacklist and not (allowMine and castByPlayer) then return end
             end
             if filterMount then
                 if isMountAura(spellId) then return true end
@@ -355,7 +355,7 @@ local function ShouldShowBuff(unit, auraData, frameType)
             local filterOnlyMe = BetterBlizzFramesDB["targetdeBuffFilterOnlyMe"] and (caster == "player" or (caster == "pet" and UnitIsUnit(caster, "pet")))
             if shouldBlacklist then
                 local isInBlacklist, allowMine = isInBlacklist(spellName, spellId)
-                if isInBlacklist and not allowMine then return end
+                if isInBlacklist and not (allowMine and castByPlayer) then return end
             end
             if not castByPlayer and onlyMine then return end
             if filterWatchlist or filterLessMinite or filterBlizzard or filterOnlyMe or isImportant or isPandemic or isEnlarged or isCompacted then return true, isImportant, isPandemic, isEnlarged, isCompacted, auraColor end
@@ -377,7 +377,7 @@ local function ShouldShowBuff(unit, auraData, frameType)
             local filterOnlyMe = BetterBlizzFramesDB["focusBuffFilterOnlyMe"] and isTargetFriendly and (caster == "player" or (caster == "pet" and UnitIsUnit(caster, "pet")))
             if shouldBlacklist then
                 local isInBlacklist, allowMine = isInBlacklist(spellName, spellId)
-                if isInBlacklist and not allowMine then return end
+                if isInBlacklist and not (allowMine and castByPlayer) then return end
             end
             if filterMount then
                 if isMountAura(spellId) then return true end
@@ -398,7 +398,7 @@ local function ShouldShowBuff(unit, auraData, frameType)
             local filterOnlyMe = BetterBlizzFramesDB["focusdeBuffFilterOnlyMe"] and (caster == "player" or (caster == "pet" and UnitIsUnit(caster, "pet")))
             if shouldBlacklist then
                 local isInBlacklist, allowMine = isInBlacklist(spellName, spellId)
-                if isInBlacklist and not allowMine then return end
+                if isInBlacklist and not (allowMine and castByPlayer) then return end
             end
             if not castByPlayer and onlyMine then return end
             if filterWatchlist or filterLessMinite or filterBlizzard or filterOnlyMe or isImportant or isPandemic or isEnlarged or isCompacted then return true, isImportant, isPandemic, isEnlarged, isCompacted, auraColor end
@@ -418,7 +418,7 @@ local function ShouldShowBuff(unit, auraData, frameType)
                 local filterLessMinite = BetterBlizzFramesDB["PlayerAuraFrameBuffFilterLessMinite"] and (duration < 61 and duration ~= 0 and expirationTime ~= 0)
                 if shouldBlacklist then
                     local isInBlacklist, allowMine = isInBlacklist(spellName, spellId)
-                    if isInBlacklist and not allowMine then return end
+                    if isInBlacklist and not (allowMine and castByPlayer) then return end
                 end
                 if filterMount then
                     if isMountAura(spellId) then return true end
@@ -437,7 +437,7 @@ local function ShouldShowBuff(unit, auraData, frameType)
                 local filterLessMinite = BetterBlizzFramesDB["PlayerAuraFramedeBuffFilterLessMinite"] and (duration < 61 and duration ~= 0 and expirationTime ~= 0)
                 if shouldBlacklist then
                     local isInBlacklist, allowMine = isInBlacklist(spellName, spellId)
-                    if isInBlacklist and not allowMine then return end
+                    if isInBlacklist and not (allowMine and castByPlayer) then return end
                 end
                 if filterWatchlist or filterLessMinite or isImportant then return true, isImportant, isPandemic, isEnlarged, isCompacted, auraColor end
                 if not BetterBlizzFramesDB["PlayerAuraFramedeBuffFilterLessMinite"] and not BetterBlizzFramesDB["PlayerAuraFramedeBuffFilterWatchList"] then
@@ -1045,14 +1045,23 @@ local function AdjustAuras(self, frameType)
                         if IsShiftKeyDown() and IsAltKeyDown() then
                             local spellName, _, icon = BBF.TWWGetSpellInfo(aura.spellId)
                             local spellId = tostring(aura.spellId)
-                            local iconString = "|T" .. icon .. ":16:16:0:0|t" -- Format the icon for display
+                            local iconString = "|T" .. icon .. ":16:16:0:0|t"
 
                             if button == "LeftButton" then
                                 BBF.auraWhitelist(aura.spellId)
-                                print("|A:gmchat-icon-blizz:16:16|a Better|cff00c0ffBlizz|rFrames: " .. iconString .. " " .. spellName .. " (" .. spellId .. ") was added to |cff00ff00whitelist|r.")
+                                print("|A:gmchat-icon-blizz:16:16|a Better|cff00c0ffBlizz|rFrames: " .. iconString .. " " .. spellName .. " (" .. spellId .. ") added to |cff00ff00whitelist|r.")
                             elseif button == "RightButton" then
                                 BBF.auraBlacklist(aura.spellId)
-                                print("|A:gmchat-icon-blizz:16:16|a Better|cff00c0ffBlizz|rFrames: " .. iconString .. " " .. spellName .. " (" .. spellId .. ") was added to |cffff0000blacklist|r.")
+                                print("|A:gmchat-icon-blizz:16:16|a Better|cff00c0ffBlizz|rFrames: " .. iconString .. " " .. spellName .. " (" .. spellId .. ") added to |cffff0000blacklist|r.")
+                            end
+                        elseif IsControlKeyDown() and IsAltKeyDown() then
+                            local spellName, _, icon = BBF.TWWGetSpellInfo(aura.spellId)
+                            local spellId = tostring(aura.spellId)
+                            local iconString = "|T" .. icon .. ":16:16:0:0|t"
+
+                            if button == "RightButton" then
+                                BBF.auraBlacklist(aura.spellId, true)
+                                print("|A:gmchat-icon-blizz:16:16|a Better|cff00c0ffBlizz|rFrames: " .. iconString .. " " .. spellName .. " (" .. spellId .. ") added to |cffff0000blacklist|r with tag.")
                             end
                         end
                     end)
@@ -1719,13 +1728,13 @@ local function PersonalBuffFrameFilterAndGrid(self)
                     auraFrame:ClearAllPoints();
                     if addIconsToRight then
                         if addIconsToTop then
-                            auraFrame:SetPoint("BOTTOMLEFT", BuffFrame, "BOTTOMLEFT", xOffset + 15, -yOffset);
+                            auraFrame:SetPoint("BOTTOMLEFT", BuffFrame, "BOTTOMLEFT", xOffset + 15, yOffset);
                         else
                             auraFrame:SetPoint("TOPLEFT", BuffFrame, "TOPLEFT", xOffset + 15, -yOffset);
                         end
                     else
                         if addIconsToTop then
-                            auraFrame:SetPoint("BOTTOMRIGHT", BuffFrame, "BOTTOMRIGHT", -xOffset - 15, -yOffset);
+                            auraFrame:SetPoint("BOTTOMRIGHT", BuffFrame, "BOTTOMRIGHT", -xOffset - 15, yOffset);
                         else
                             auraFrame:SetPoint("TOPRIGHT", BuffFrame, "TOPRIGHT", -xOffset - 15, -yOffset);
                         end
@@ -1738,14 +1747,23 @@ local function PersonalBuffFrameFilterAndGrid(self)
                             if IsShiftKeyDown() and IsAltKeyDown() then
                                 local spellName, _, icon = BBF.TWWGetSpellInfo(auraFrame.spellId)
                                 local spellId = tostring(auraFrame.spellId)
-                                local iconString = "|T" .. icon .. ":16:16:0:0|t" -- Format the icon for display
+                                local iconString = "|T" .. icon .. ":16:16:0:0|t"
 
                                 if button == "LeftButton" then
                                     BBF.auraWhitelist(auraFrame.spellId)
-                                    print("|A:gmchat-icon-blizz:16:16|a Better|cff00c0ffBlizz|rFrames: " .. iconString .. " " .. spellName .. " (" .. spellId .. ") was added to |cff00ff00whitelist|r.")
+                                    print("|A:gmchat-icon-blizz:16:16|a Better|cff00c0ffBlizz|rFrames: " .. iconString .. " " .. spellName .. " (" .. spellId .. ") added to |cff00ff00whitelist|r.")
                                 elseif button == "RightButton" then
                                     BBF.auraBlacklist(auraFrame.spellId)
-                                    print("|A:gmchat-icon-blizz:16:16|a Better|cff00c0ffBlizz|rFrames: " .. iconString .. " " .. spellName .. " (" .. spellId .. ") was added to |cffff0000blacklist|r.")
+                                    print("|A:gmchat-icon-blizz:16:16|a Better|cff00c0ffBlizz|rFrames: " .. iconString .. " " .. spellName .. " (" .. spellId .. ") added to |cffff0000blacklist|r.")
+                                end
+                            elseif IsControlKeyDown() and IsAltKeyDown() then
+                                local spellName, _, icon = BBF.TWWGetSpellInfo(auraFrame.spellId)
+                                local spellId = tostring(auraFrame.spellId)
+                                local iconString = "|T" .. icon .. ":16:16:0:0|t"
+    
+                                if button == "RightButton" then
+                                    BBF.auraBlacklist(auraFrame.spellId, true)
+                                    print("|A:gmchat-icon-blizz:16:16|a Better|cff00c0ffBlizz|rFrames: " .. iconString .. " " .. spellName .. " (" .. spellId .. ") added to |cffff0000blacklist|r with tag.")
                                 end
                             end
                         end)
@@ -2034,14 +2052,23 @@ local function PersonalDebuffFrameFilterAndGrid(self)
                             if IsShiftKeyDown() and IsAltKeyDown() then
                                 local spellName, _, icon = BBF.TWWGetSpellInfo(auraFrame.spellId)
                                 local spellId = tostring(auraFrame.spellId)
-                                local iconString = "|T" .. icon .. ":16:16:0:0|t" -- Format the icon for display
+                                local iconString = "|T" .. icon .. ":16:16:0:0|t"
 
                                 if button == "LeftButton" then
                                     BBF.auraWhitelist(auraFrame.spellId)
-                                    print("|A:gmchat-icon-blizz:16:16|a Better|cff00c0ffBlizz|rFrames: " .. iconString .. " " .. spellName .. " (" .. spellId .. ") was added to |cff00ff00whitelist|r.")
+                                    print("|A:gmchat-icon-blizz:16:16|a Better|cff00c0ffBlizz|rFrames: " .. iconString .. " " .. spellName .. " (" .. spellId .. ") added to |cff00ff00whitelist|r.")
                                 elseif button == "RightButton" then
                                     BBF.auraBlacklist(auraFrame.spellId)
-                                    print("|A:gmchat-icon-blizz:16:16|a Better|cff00c0ffBlizz|rFrames: " .. iconString .. " " .. spellName .. " (" .. spellId .. ") was added to |cffff0000blacklist|r.")
+                                    print("|A:gmchat-icon-blizz:16:16|a Better|cff00c0ffBlizz|rFrames: " .. iconString .. " " .. spellName .. " (" .. spellId .. ") added to |cffff0000blacklist|r.")
+                                end
+                            elseif IsControlKeyDown() and IsAltKeyDown() then
+                                local spellName, _, icon = BBF.TWWGetSpellInfo(auraFrame.spellId)
+                                local spellId = tostring(auraFrame.spellId)
+                                local iconString = "|T" .. icon .. ":16:16:0:0|t"
+    
+                                if button == "RightButton" then
+                                    BBF.auraBlacklist(auraFrame.spellId, true)
+                                    print("|A:gmchat-icon-blizz:16:16|a Better|cff00c0ffBlizz|rFrames: " .. iconString .. " " .. spellName .. " (" .. spellId .. ") added to |cffff0000blacklist|r with tag.")
                                 end
                             end
                         end)
@@ -2051,11 +2078,18 @@ local function PersonalDebuffFrameFilterAndGrid(self)
                     auraFrame:Show();
                     auraFrame:ClearAllPoints();
                     if addIconsToRight then
-                        auraFrame:SetPoint("TOPLEFT", DebuffFrame, "TOPLEFT", xOffset, -yOffset);
+                        if addIconsToTop then
+                            auraFrame:SetPoint("BOTTOMLEFT", DebuffFrame, "BOTTOMLEFT", xOffset, yOffset);
+                        else
+                            auraFrame:SetPoint("TOPLEFT", DebuffFrame, "TOPLEFT", xOffset, -yOffset);
+                        end
                     else
-                        auraFrame:SetPoint("TOPRIGHT", DebuffFrame, "TOPRIGHT", -xOffset, -yOffset);
+                        if addIconsToTop then
+                            auraFrame:SetPoint("BOTTOMRIGHT", DebuffFrame, "BOTTOMRIGHT", -xOffset, yOffset);
+                        else
+                            auraFrame:SetPoint("TOPRIGHT", DebuffFrame, "TOPRIGHT", -xOffset, -yOffset);
+                        end
                     end
-                    --print(auraFrame:GetSize())
 
                     -- Update column and row counters
                     currentCol = currentCol + 1;